+++ modules/tinymce/src/core/main/ts/file/ImageScanner.ts
@@ -1,8 +1,8 @@
-import { Arr, Fun, Strings } from '@ephox/katamari';
+import { Arr, Fun, Obj, Strings, Type } from '@ephox/katamari';
 
 import Env from '../api/Env';
 import { BlobCache, BlobInfo } from '../api/file/BlobCache';
-import * as Conversions from './Conversions';
+import { imageToBlobInfo } from './BlobCacheUtils';
 import { UploadStatus } from './UploadStatus';
 
 export interface BlobInfoImagePair {
@@ -11,7 +11,7 @@ export interface BlobInfoImagePair {
 }
 
 export interface ImageScanner {
-  findAll: (elm: HTMLElement, predicate?: (img: HTMLImageElement) => boolean) => Promise<BlobInfoImagePair[]>;
+  findAll: (elm: HTMLElement, predicate?: (img: HTMLImageElement) => boolean) => Promise<Array<BlobInfoImagePair | string>>;
 }
 
 /**
@@ -20,61 +20,6 @@ export interface ImageScanner {
  * @private
  * @class tinymce.file.ImageScanner
  */
-
-let count = 0;
-
-export const uniqueId = (prefix?: string): string => {
-  return (prefix || 'blobid') + (count++);
-};
-
-const imageToBlobInfo = (blobCache: BlobCache, img: HTMLImageElement): Promise<BlobInfoImagePair> => {
-  let base64, blobInfo;
-
-  if (img.src.indexOf('blob:') === 0) {
-    blobInfo = blobCache.getByUri(img.src);
-
-    if (blobInfo) {
-      return Promise.resolve({
-        image: img,
-        blobInfo
-      });
-    } else {
-      return Conversions.uriToBlob(img.src).then((blob) => {
-        return Conversions.blobToDataUri(blob).then((dataUri) => {
-          base64 = Conversions.parseDataUri(dataUri).data;
-          blobInfo = blobCache.create(uniqueId(), blob, base64);
-          blobCache.add(blobInfo);
-
-          return {
-            image: img,
-            blobInfo
-          };
-        });
-      });
-    }
-  }
-
-  const { data, type } = Conversions.parseDataUri(img.src);
-  base64 = data;
-  blobInfo = blobCache.getByData(base64, type);
-
-  if (blobInfo) {
-    return Promise.resolve({
-      image: img,
-      blobInfo
-    });
-  } else {
-    return Conversions.uriToBlob(img.src).then((blob) => {
-      blobInfo = blobCache.create(uniqueId(), blob, base64);
-      blobCache.add(blobInfo);
-
-      return {
-        image: img,
-        blobInfo
-      };
-    });
-  }
-};
 
 const getAllImages = (elm: HTMLElement): HTMLImageElement[] => {
   return elm ? Arr.from(elm.getElementsByTagName('img')) : [];
@@ -111,34 +56,35 @@ export const ImageScanner = (uploadStatus: UploadStatus, blobCache: BlobCache): 
     });
 
     const promises = Arr.map(images, (img): Promise<BlobInfoImagePair> => {
-      if (cachedPromises[img.src] !== undefined) {
+      const imageSrc = img.src;
+
+      if (Obj.has(cachedPromises, imageSrc)) {
         // Since the cached promise will return the cached image
         // We need to wrap it and resolve with the actual image
-        return new Promise((resolve) => {
-          cachedPromises[img.src].then((imageInfo) => {
-            if (typeof imageInfo === 'string') { // error apparently
-              return imageInfo;
-            }
-            resolve({
+        return cachedPromises[imageSrc].then((imageInfo) => {
+          if (Type.isString(imageInfo)) { // error apparently
+            return imageInfo;
+          } else {
+            return {
               image: img,
               blobInfo: imageInfo.blobInfo
-            });
+            };
+          }
+        });
+      } else {
+        const newPromise = imageToBlobInfo(blobCache, imageSrc)
+          .then((blobInfo) => {
+            delete cachedPromises[imageSrc];
+            return { image: img, blobInfo };
+          }).catch((error) => {
+            delete cachedPromises[imageSrc];
+            return error;
           });
-        });
+
+        cachedPromises[imageSrc] = newPromise;
+
+        return newPromise;
       }
-
-      const newPromise = imageToBlobInfo(blobCache, img)
-        .then((result) => {
-          delete cachedPromises[result.image.src];
-          return result;
-        }).catch((error) => {
-          delete cachedPromises[img.src];
-          return error;
-        });
-
-      cachedPromises[img.src] = newPromise;
-
-      return newPromise;
     });
 
     return Promise.all(promises);