+++ modules/tinymce/src/core/test/ts/browser/file/ImageScannerTest.ts
@@ -5,7 +5,7 @@ import { assert } from 'chai';
 import Env from 'tinymce/core/api/Env';
 import { BlobCache } from 'tinymce/core/api/file/BlobCache';
 import * as Conversions from 'tinymce/core/file/Conversions';
-import { ImageScanner } from 'tinymce/core/file/ImageScanner';
+import { BlobInfoImagePair, ImageScanner } from 'tinymce/core/file/ImageScanner';
 import { UploadStatus } from 'tinymce/core/file/UploadStatus';
 
 import * as ViewBlock from '../../module/test/ViewBlock';
@@ -13,6 +13,8 @@ import * as ViewBlock from '../../module/test/ViewBlock';
 describe('browser.tinymce.core.file.ImageScannerTest', () => {
   const viewBlock = ViewBlock.bddSetup();
   const base64Src = 'data:image/gif;base64,R0lGODlhAQABAIAAAAAAAAAAACH5BAAAAAAALAAAAAABAAEAAAICTAEAOw==';
+  const svg = `<svg xmlns="http://www.w3.org/2000/svg" height="100" width="100"><circle cx="50" cy="50" r="40" stroke="black" stroke-width="3" fill="red"></svg>`;
+  const encodedSrc = 'data:image/svg+xml,' + encodeURIComponent(svg);
   const invalidBlobUriSrc = 'blob:70BE8432-BA4D-4787-9AB9-86563351FBF7';
   let blobUriSrc;
 
@@ -28,6 +30,7 @@ describe('browser.tinymce.core.file.ImageScannerTest', () => {
     viewBlock.update(
       '<img src="' + base64Src + '">' +
       '<img src="' + blobUriSrc + '">' +
+      '<img src="' + encodedSrc + '">' +
       '<img src="' + Env.transparentSrc + '">' +
       '<img src="' + base64Src + '" data-mce-bogus="1">' +
       '<img src="' + base64Src + '" data-mce-placeholder="1">' +
@@ -35,11 +38,14 @@ describe('browser.tinymce.core.file.ImageScannerTest', () => {
     );
 
     return imageScanner.findAll(viewBlock.get()).then((result) => {
-      const blobInfo = result[0].blobInfo;
-      assert.lengthOf(result, 3);
+      assert.lengthOf(result, 4);
+      const base64ImageResult = result[0] as BlobInfoImagePair;
+      const encodedImageResult = result[2] as BlobInfoImagePair;
       assert.typeOf(result[result.length - 1], 'string', 'Last item is not the image, but error message.');
-      assert.equal('data:image/gif;base64,' + blobInfo.base64(), base64Src);
-      LegacyUnit.equalDom(result[0].image, viewBlock.get().firstChild);
+      assert.equal('data:image/gif;base64,' + base64ImageResult.blobInfo.base64(), base64Src);
+      LegacyUnit.equalDom(base64ImageResult.image, viewBlock.get().firstChild);
+      assert.equal('data:image/svg+xml;base64,' + encodedImageResult.blobInfo.base64(), 'data:image/svg+xml;base64,' + btoa(svg));
+      LegacyUnit.equalDom(encodedImageResult.image, viewBlock.get().childNodes.item(2));
     });
   });
 
@@ -57,8 +63,9 @@ describe('browser.tinymce.core.file.ImageScannerTest', () => {
 
     return imageScanner.findAll(viewBlock.get(), predicate).then((result) => {
       assert.lengthOf(result, 1);
-      assert.equal('data:image/gif;base64,' + result[0].blobInfo.base64(), base64Src);
-      LegacyUnit.equalDom(result[0].image, viewBlock.get().firstChild);
+      const firstResult = result[0] as BlobInfoImagePair;
+      assert.equal('data:image/gif;base64,' + firstResult.blobInfo.base64(), base64Src);
+      LegacyUnit.equalDom(firstResult.image, viewBlock.get().firstChild);
     });
   });
 });